Rename AsyncRepeater and AsyncRunner to FrameAsync~.
