+++ Assets/Tests/FrameAsyncRunnerTest.cs
@@ -8,7 +8,7 @@ using TSAsyncTools;
 
 namespace Tests
 {
-    public class AsyncRunnerTest
+    public class FrameAsyncRunnerTest
     {
         [UnityTest]
         public IEnumerator Runをコールした後にActionが実施されること()
@@ -17,7 +17,7 @@ namespace Tests
             int callCount = 0;
             Action action = () => { callCount += 1; };
             var cts = new CancellationTokenSource();
-            var target = new AsyncRunner(action, 0, PlayerLoopTiming.PreUpdate, cts.Token);
+            var target = new FrameAsyncRunner(action, 0, PlayerLoopTiming.PreUpdate, cts.Token);
 
             // when
             target.Run();
@@ -34,7 +34,7 @@ namespace Tests
             int callCount = 0;
             Action action = () => { callCount += 1; };
             var cts = new CancellationTokenSource();
-            var target = new AsyncRunner(action, 0, PlayerLoopTiming.PreUpdate, cts.Token);
+            var target = new FrameAsyncRunner(action, 0, PlayerLoopTiming.PreUpdate, cts.Token);
 
             // when
             target.Run();
@@ -52,7 +52,7 @@ namespace Tests
             int callCount = 0;
             Action action = () => { callCount += 1; };
             var cts = new CancellationTokenSource();
-            var target = new AsyncRunner(action, 0, PlayerLoopTiming.PreUpdate, cts.Token);
+            var target = new FrameAsyncRunner(action, 0, PlayerLoopTiming.PreUpdate, cts.Token);
 
             // when
             cts.Cancel();
@@ -70,7 +70,7 @@ namespace Tests
             int callCount = 0;
             Action action = () => { callCount += 1; };
             var cts = new CancellationTokenSource();
-            var target = new AsyncRunner(action, 0, PlayerLoopTiming.PreUpdate, cts.Token);
+            var target = new FrameAsyncRunner(action, 0, PlayerLoopTiming.PreUpdate, cts.Token);
 
             // when
             target.Run();
@@ -83,13 +83,13 @@ namespace Tests
         }
 
         [UnityTest]
-        public IEnumerator フレームごとにRunをコールした場合にコールした回数Actionが実施されること()
+        public IEnumerator Action完了後に再度Runをコールした場合に再びActionが実施されること()
         {
             // setup
             int callCount = 0;
             Action action = () => { callCount += 1; };
             var cts = new CancellationTokenSource();
-            var target = new AsyncRunner(action, 0, PlayerLoopTiming.PreUpdate, cts.Token);
+            var target = new FrameAsyncRunner(action, 0, PlayerLoopTiming.PreUpdate, cts.Token);
 
             // when
             target.Run();
@@ -110,12 +110,73 @@ namespace Tests
             int callCount = 0;
             Action action = () => { callCount += 1; };
             var cts = new CancellationTokenSource();
-            var target = new AsyncRunner(action, 0, PlayerLoopTiming.PreUpdate, cts.Token);
+            var target = new FrameAsyncRunner(action, 0, PlayerLoopTiming.PreUpdate, cts.Token);
 
             // when
             target.Run();
             for (int i = 0; i < 30; ++i) { yield return null; }
             target.Run();
+            yield return null;
+
+            // then
+            Assert.AreEqual(2, callCount);
+        }
+
+        [UnityTest]
+        public IEnumerator 遅延フレームを指定した場合に遅延フレーム後にActionが実施されること()
+        {
+            // setup
+            int callCount = 0;
+            Action action = () => { callCount += 1; };
+            var cts = new CancellationTokenSource();
+            var target = new FrameAsyncRunner(action, 3, PlayerLoopTiming.PreUpdate, cts.Token);
+
+            // when
+            target.Run();
+            yield return null;
+            yield return null;
+            yield return null;
+
+            // then
+            Assert.AreEqual(1, callCount);
+        }
+
+        [UnityTest]
+        public IEnumerator 遅延フレームを指定した場合に遅延フレーム前にActionが実施されないこと()
+        {
+            // setup
+            int callCount = 0;
+            Action action = () => { callCount += 1; };
+            var cts = new CancellationTokenSource();
+            var target = new FrameAsyncRunner(action, 3, PlayerLoopTiming.PreUpdate, cts.Token);
+
+            // when
+            target.Run();
+            yield return null;
+            yield return null;
+
+            // then
+            Assert.AreEqual(0, callCount);
+        }
+
+
+        [UnityTest]
+        public IEnumerator 遅延フレームを指定してもAction完了後に再度Runをコールした場合に再びActionが実施されること()
+        {
+            // setup
+            int callCount = 0;
+            Action action = () => { callCount += 1; };
+            var cts = new CancellationTokenSource();
+            var target = new FrameAsyncRunner(action, 3, PlayerLoopTiming.PreUpdate, cts.Token);
+
+            // when
+            target.Run();
+            yield return null;
+            yield return null;
+            yield return null;
+            target.Run();
+            yield return null;
+            yield return null;
             yield return null;
 
             // then
